# addOns/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Bus Api",
    "version": "v1"
  },
  "paths": {
    "/api/Admin/lines": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Line"
                  },
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getLines"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "addLine"
      }
    },
    "/api/Admin/lines/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Line"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getLine"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Line"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "editLine"
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "deleteLine"
      }
    },
    "/api/Admin/stations": {
      "get": {
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  },
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getStations"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "addStation"
      }
    },
    "/api/Admin/stations/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getStation"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "editStation"
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "deleteStation"
      }
    },
    "/api/Admin/timetables/{lineId}/{day}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "x-enumNames": [
              "Sunday",
              "Monday",
              "Tuesday",
              "Wednesday",
              "Thursday",
              "Friday",
              "Saturday"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timetable"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getTimeTable"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "lineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          },
          {
            "name": "day",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "x-enumNames": [
              "Sunday",
              "Monday",
              "Tuesday",
              "Wednesday",
              "Thursday",
              "Friday",
              "Saturday"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Timetable"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "editTimeTable"
      }
    },
    "/api/Admin/pricelists/{ticketType}/{passengerType}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "ticketType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TicketType"
            },
            "x-enumNames": [
              "TimeTicket",
              "DailyTicket",
              "MonthlyTicket",
              "YearlyTicket"
            ]
          },
          {
            "name": "passengerType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PassengerType"
            },
            "x-enumNames": [
              "Regular",
              "Student",
              "Pensioner"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pricelist"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getTicketPrice"
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "ticketType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/TicketType"
            },
            "x-enumNames": [
              "TimeTicket",
              "DailyTicket",
              "MonthlyTicket",
              "YearlyTicket"
            ]
          },
          {
            "name": "passengerType",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PassengerType"
            },
            "x-enumNames": [
              "Regular",
              "Student",
              "Pensioner"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Pricelist"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Pricelist"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Pricelist"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Pricelist"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean",
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "setTicketPrice"
      }
    },
...
    "/api/Lines": {
      "get": {
        "tags": [
          "Lines"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Line"
                  },
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getLines"
      }
    },
...
    "/api/Lines/{id}": {
      "get": {
        "tags": [
          "Lines"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "x-enumNames": [ ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Line"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getLine"
      }
    },
...
    "/api/Stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  },
                  "x-enumNames": [ ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApiResponse"
                }
              }
            }
          }
        },
        "x-operation-name": "getStations"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Line": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "x-enumNames": [ ]
          },
          "name": {
            "type": "string",
            "nullable": true,
            "x-enumNames": [ ]
          },
          "color": {
            "type": "string",
            "nullable": true,
            "x-enumNames": [ ]
          },
          "lineType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/LineType"
              }
            ]
          },
          "waypoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GeoLocation"
            },
            "nullable": true,
            "x-enumNames": [ ]
          },
          "stations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StationLine"
            },
            "nullable": true,
            "x-enumNames": [ ]
          }
        },
        "additionalProperties": false,
        "x-enumNames": [ ]
      },
...
      "TicketType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "TimeTicket",
          "DailyTicket",
          "MonthlyTicket",
          "YearlyTicket"
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}